+++ src/components/shotlist/TakeList.jsx
@@ -1,5 +1,6 @@
-import { Fragment } from 'react';
+import { Fragment, useEffect, useState } from 'react';
 
+import { useGAPITokenHandler } from '@/lib/googleapi';
 import { useTakeIds } from '@/stores/DocumentStoreContext';
 
 import { NewTake, TakeEntry } from './TakeEntry';
@@ -12,8 +13,17 @@ import { NewTake, TakeEntry } from './TakeEntry';
  */
 export default function TakeList({ documentId, sceneId, shotId }) {
   const takeIds = useTakeIds(documentId, shotId);
+  const [cloudExportable, setCloudExportable] = useState(false);
+  const handleToken = useGAPITokenHandler();
+
+  useEffect(() => {
+    if (!handleToken((token) => setCloudExportable(Boolean(token)))) {
+      setCloudExportable(false);
+    }
+  }, [handleToken, setCloudExportable]);
+
   return (
-    <ul className="mx-8">
+    <ul>
       <NewTake documentId={documentId} shotId={shotId} />
       {takeIds
         .slice()
@@ -25,6 +35,7 @@ export default function TakeList({ documentId, sceneId, shotId }) {
               sceneId={sceneId}
               shotId={shotId}
               takeId={takeId}
+              cloudExportable={cloudExportable}
             />
           </Fragment>
         ))}